fix(summary): wire configured intent into diagnostic callers
Refs #294
