+++ scripts/check-yesterday-reader-summary-evidence-replay.ts
@@ -1,3 +1,4 @@
+import { checkConfiguredInterestReader, requireFreshCheckSelection } from "./lib/check-configured-interest-reader";
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
 import { dirname } from "node:path";
 
@@ -26,6 +27,7 @@ import { FixedClock } from "@social-monitor/shared-kernel";
 
 type Report = {
   readonly schemaVersion: 1;
+  readonly selectionAuthority: "current_monitoring_configuration";
   readonly artifactFormat: "yesterday-reader-summary-evidence-replay-v1";
   readonly collectionDate: string;
   readonly generatedBy: string;
@@ -83,7 +85,7 @@ type Report = {
 const { collectionDate } = collectionDateOptionOrDefault("2026-07-03");
 const update = process.argv.includes("--update");
 const allowDirtyCollection = process.argv.includes("--allow-dirty-collection");
-const outputPath = "ops/evals/yesterday-reader-summary-evidence-replay.v1.json";
+const outputPath = "ops/evals/yesterday-reader-summary-evidence-replay.fresh.v1.json";
 const maxEvidenceItems = 40;
 const primarySources = ["reddit", "x-twitter"];
 const localDatabaseUrl = yesterdaySocialQualityDatabaseUrl();
@@ -91,6 +93,7 @@ const localDatabaseUrl = yesterdaySocialQualityDatabaseUrl();
 void main();
 
 async function main(): Promise<void> {
+  requireFreshCheckSelection(process.argv.slice(2));
   const report = await tryBuildReport();
 
   if (report === undefined) {
@@ -119,14 +122,14 @@ async function main(): Promise<void> {
 
   if (!existsSync(outputPath)) {
     throw new Error(
-      `${outputPath} is missing. Run npm run check:yesterday-reader-summary-evidence-replay -- --update`,
+      `${outputPath} is missing. Run npm run check:yesterday-reader-summary-evidence-replay -- --fresh-selection --update`,
     );
   }
 
   const expected = normalizeLineEndings(readFileSync(outputPath, "utf8"));
   if (expected !== serialized) {
     throw new Error(
-      `${outputPath} is stale. Run npm run check:yesterday-reader-summary-evidence-replay -- --update`,
+      `${outputPath} is stale. Run npm run check:yesterday-reader-summary-evidence-replay -- --fresh-selection --update`,
     );
   }
 
@@ -158,6 +161,8 @@ async function tryBuildReport(): Promise<Report | undefined> {
       feedItems,
       new InMemoryUserRelevanceProfileRepository(),
       new FixedClock(new Date(`${collectionDate}T23:59:59.000Z`)),
+      undefined, undefined, undefined, undefined, undefined,
+      checkConfiguredInterestReader(localDatabaseUrl),
     );
     const selector = new RelevanceReaderSummaryEvidenceSelector(
       rankFeedItems,
@@ -256,9 +261,10 @@ async function tryBuildReport(): Promise<Report | undefined> {
     };
     const reportWithoutSecretGate = {
       schemaVersion: 1,
+      selectionAuthority: "current_monitoring_configuration",
       artifactFormat: "yesterday-reader-summary-evidence-replay-v1",
       collectionDate,
-      generatedBy: "npm run check:yesterday-reader-summary-evidence-replay",
+      generatedBy: "npm run check:yesterday-reader-summary-evidence-replay -- --fresh-selection",
       model: {
         liveNetwork: false,
         replayTarget: "workspace-reader-summary-evidence",
@@ -321,9 +327,7 @@ function validateExistingReport(): void {
     throw new Error(`${outputPath} failed existing artifact validation`);
   }
 
-  console.log(
-    `Yesterday reader summary evidence replay artifact OK (${report.collectionDate}; local source unavailable)`,
-  );
+  throw new Error("Fresh selection unavailable; stored replay report is not current configuration evidence");
 }
 
 function countBy<TValue>(
